Move FormXCubit to bloc directory
