+++ lib/flutter_formx.dart
@@ -1,4 +1,4 @@
-export 'src/form/cubit/formx_cubit.dart';
+export 'src/form/bloc/formx_cubit.dart';
 export 'src/form/formx_field.dart';
 export 'src/form/mobx/formx_mobx.dart';
 export 'src/validator/required_field_validator.dart';
